requirements.txt
asgiref==3.11.0
dj-database-url==3.0.1
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
greenlet==3.2.4
playwright==1.55.0
psycopg2-binary==2.9.11
pyee==13.0.0
sqlparse==0.5.4
typing_extensions==4.15.0
whitenoise==6.11.0
